# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 2026 Taiwan Balloon Carnival API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/checkin/lookup/{token}:
    get:
      operationId: adminLookupCheckin
      tags: [admin]
      summary: Look up a registration by QR token (admin)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Registration found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinLookupResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/checkin/{token}:
    post:
      operationId: adminPerformCheckin
      tags: [admin]
      summary: Mark a registration as checked-in (admin)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinLookupResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Already checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinLookupResponse"
  /admin/email-templates:
    get:
      operationId: adminListEmailTemplates
      tags: [admin]
      summary: List all email templates (admin)
      responses:
        "200":
          description: Email templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailTemplate"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /admin/email-templates/{key}:
    put:
      operationId: adminUpdateEmailTemplate
      tags: [admin]
      summary: Update an email template (admin)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmailTemplateBody"
      responses:
        "200":
          description: Template updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Template not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/email-templates/{key}/test:
    post:
      operationId: adminSendTestEmail
      tags: [admin]
      summary: Send a test email using the template (admin)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendTestEmailBody"
      responses:
        "200":
          description: Test email sent (or logged)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendTestEmailResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Registration:
      type: object
      required: [id, parentName, phone, ticketCount, eventDate, paymentStatus, createdAt]
      properties:
        id:
          type: number
        parentName:
          type: string
        phone:
          type: string
        email:
          type: [string, 'null']
        ticketCount:
          type: number
        eventDate:
          type: string
          format: date
        ticketType:
          type: [string, 'null']
        amount:
          type: [number, 'null']
        paymentMethod:
          type: [string, 'null']
        paymentStatus:
          type: string
        paymentRef:
          type: [string, 'null']
        qrToken:
          type: [string, 'null']
        checkedInAt:
          type: [string, 'null']
          format: date-time
        confirmationEmailSentAt:
          type: [string, 'null']
          format: date-time
        weekReminderSentAt:
          type: [string, 'null']
          format: date-time
        dayReminderSentAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    EmailTemplate:
      type: object
      required: [id, key, subject, body, updatedAt]
      properties:
        id:
          type: number
        key:
          type: string
        subject:
          type: string
        body:
          type: string
        updatedAt:
          type: string
          format: date-time
    UpdateEmailTemplateBody:
      type: object
      required: [subject, body]
      properties:
        subject:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
    SendTestEmailBody:
      type: object
      required: [recipient]
      properties:
        recipient:
          type: string
          format: email
    SendTestEmailResponse:
      type: object
      required: [delivered, mode]
      properties:
        delivered:
          type: boolean
        mode:
          type: string
          enum: [resend, console]
        message:
          type: string
    CheckinLookupResponse:
      type: object
      required: [registration, alreadyCheckedIn]
      properties:
        registration:
          $ref: "#/components/schemas/Registration"
        alreadyCheckedIn:
          type: boolean
...